Offer the user the option to configure an AutoFill profile.
When the user selects a form field that is part of a form that can
be autofilled, inform the embedder that this is the case to give
them the opportunity to configure a profile if they the user has
not done so already. A Message is also provided
to the embedder that when sent will trigger the form to be filled
with the profile that the embedder set up.

Change-Id: Ica995e5b1ed92a3ec3e356401b261740d9f90e57