Use SearchView in SearchDialog for legacy apps using the old model.
Bug: 3140925

Removed all the code from SearchDialog that is already in SearchView
and moved more stuff into SearchView.

SearchDialog layout changed to be more like a Holo.Light themed action mode.
Search field is right justified and limited to 500dip.

Zero query dropdown added to SearchView (causing some problems with Gmail,
but that may be an existing issue that's only now showing up due to this change)

Holo.Light search views assets still need to be updated, as the contrast is too low.